GEOMESA-2639 Additional package standardization
* org.locationtech.geomesa.jobs.accumulo -> org.locationtech.geomesa.accumulo.jobs
* org.locationtech.geomesa.spark.accumulo -> org.locationtech.geomesa.accumulo.spark
* org.locationtech.geomesa.spark.hbase -> org.locationtech.geomesa.hbase.spark
* Remove deprecated geomesa-convert-metrics-* modules
