+++ geomesa-accumulo/geomesa-accumulo-spark/src/main/resources/META-INF/services/org.locationtech.geomesa.spark.SpatialRDDProvider
@@ -1,1 +1,1 @@
-org.locationtech.geomesa.spark.accumulo.AccumuloSpatialRDDProvider
+org.locationtech.geomesa.accumulo.spark.AccumuloSpatialRDDProvider
