--- geomesa-convert/geomesa-convert-metrics-ganglia/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <artifactId>geomesa-convert_2.12</artifactId>
-        <groupId>org.locationtech.geomesa</groupId>
-        <version>4.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>geomesa-convert-metrics-ganglia_2.12</artifactId>
-    <name>GeoMesa Convert Metrics - Ganglia</name>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <!-- excluded due to licensing -->
-                <groupId>info.ganglia.gmetric4j</groupId>
-                <artifactId>gmetric4j</artifactId>
-                <version>1.0.7</version>
-                <scope>provided</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-convert-common_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-metrics-ganglia_${scala.binary.version}</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
