--- geomesa-convert/geomesa-convert-metrics-graphite/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <artifactId>geomesa-convert_2.12</artifactId>
-        <groupId>org.locationtech.geomesa</groupId>
-        <version>4.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>geomesa-convert-metrics-graphite_2.12</artifactId>
-    <name>GeoMesa Convert Metrics - Graphite</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-convert-common_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-metrics-graphite_${scala.binary.version}</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
