--- geomesa-convert/geomesa-convert-metrics-graphite/src/main/scala/org/locationtech/geomesa/convert/metrics/graphite/ScalaDocs.scala
@@ -1,15 +1,0 @@
-/***********************************************************************
- * Copyright (c) 2013-2022 Commonwealth Computer Research, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Apache License, Version 2.0
- * which accompanies this distribution and is available at
- * http://www.opensource.org/licenses/apache2.0.php.
- ***********************************************************************/
-
-package org.locationtech.geomesa.convert.metrics.graphite
-
-/**
- * Placeholder class to force generation of -javadocs and -sources jars for maven publishing */
-object ScalaDocs {
-
-}
