+++ geomesa-convert/pom.xml
@@ -20,9 +20,6 @@
         <module>geomesa-convert-fixedwidth</module>
         <module>geomesa-convert-jdbc</module>
         <module>geomesa-convert-json</module>
-        <module>geomesa-convert-metrics-cloudwatch</module>
-        <module>geomesa-convert-metrics-ganglia</module>
-        <module>geomesa-convert-metrics-graphite</module>
         <module>geomesa-convert-osm</module>
         <module>geomesa-convert-parquet</module>
         <module>geomesa-convert-redis-cache</module>
