+++ geomesa-hbase/geomesa-hbase-spark/src/main/resources/META-INF/services/org.locationtech.geomesa.spark.SpatialRDDProvider
@@ -1,1 +1,1 @@
-org.locationtech.geomesa.spark.hbase.HBaseSpatialRDDProvider
+org.locationtech.geomesa.hbase.spark.HBaseSpatialRDDProvider
