+++ geomesa-kafka/geomesa-kafka-utils/src/main/scala/org/locationtech/geomesa/kafka/consumer/BatchConsumer.scala
@@ -10,9 +10,9 @@ package org.locationtech.geomesa.kafka.consumer
 
 import org.apache.kafka.clients.consumer.{Consumer, ConsumerRecord}
 import org.apache.kafka.common.errors.{InterruptException, WakeupException}
-import org.locationtech.geomesa.kafka.KafkaConsumerVersions
 import org.locationtech.geomesa.kafka.consumer.BatchConsumer.BatchResult
 import org.locationtech.geomesa.kafka.consumer.ThreadedConsumer.{ConsumerErrorHandler, LogOffsetCommitCallback}
+import org.locationtech.geomesa.kafka.versions.KafkaConsumerVersions
 
 import java.time.Duration
 import java.util.Collections
